person_4/requirements.txt
# Core dependencies (aligned with Person 1/2/3 versions)
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.0
datasets>=2.16.0

# API and Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.5.0

# File parsing (PDF, DOCX support)
PyMuPDF>=1.23.0
python-docx>=0.8.11

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0
huggingface-hub>=0.20.0
safetensors>=0.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Monitoring
wandb>=0.16.0
tensorboard>=2.13.0

# Plagiarism detection (Person 2)
datasketch>=1.6.0

# GPU acceleration (needed for Person 1/3 models)
accelerate>=0.25.0
peft>=0.6.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
